✏️ Rename `core` -> `crypto`
It appears that having a crate named `core` is problematic as it
conflicts with the `no_std` `core` crate -- which prevents things like
the default `core::prelude` from being included when running doctests,
or when depending on this logic from outside the crate.

To prevent issues, this crate is now named `crypto` and will have a more
narrow scope of focus instead.