🔧 Add `cargo-deny` check to CI
`cargo-deny` is a very useful tool that can check for unwanted
transitive dependencies that have unhygenic dependency habits -- such
as depending on untrusted Git repository types, bad licenses, etc.

This integrates the tooling as part of the check, which will also
ensure that repositories are well-maintained, trusted, and use
expected licenses. As part of this change, the `Cargo.toml` file for
the `crypto` package needed to be updated to state the licenses used so
that it may be whitelisted.