✨ Add first-draft of low-level command package
When working with bare-metal software such as kernels, bootloaders, and
operating-systems -- it's required to dip into assembly from
time-to-time in order to use low-level features not exposed in higher
level tooling.

This is something that's needed for this project as well; so this
adds a first-draft mechanism for abstracting the different low-level
assembly calls between different architectures. This will likely be
changed in a future commit.

The current command being added is `halt`, for shutting down the
program.