✨ Add skeleton of Bootloader
This introduces a skeleton application that will become the bootloader.
It uses the `uefi` dependency to simply write a splash to the screen
before waiting and then shutting down.

To make building easier, a `Makefile` has been added as well, along with
a step for `run-qemu` to actually execute the bootloader.

Tracked-by: #1