🔧 Add support for reliable QEMU booting
The `Makefile` previously only supported simple execution of the
`bootloader` EFI binary by leveraging a user-supplied OVMF filesystem
location. This was not ideal, could not reliably run on other systems,
and did not support aarch64 building.

This commit now hardens the `Makefile` so that its more reliable and
easy to use. It will now retrieve the OVMF binaries for the specified
system as-needed, and has been configured so that the `run-qemu`
command will properly execute the specified architecture.

Aarch64 is now supported as a bootload target, as is x86_64.

In the process, this also makes it such that `aarch64` binaries will
build `bootaa64.efi` files, whereas `x86_64` binaries will build
`bootx86.efi` binaries. This doesn't fully address the issue that
these named binaries _shouldn't_ have to be replaced in a proper
system -- which is something that will be addressed more thoroughly
in #10.

Fixes: #9