🩹 Suppress temporary `cargo-deny` failures
The `uefi` package has triggered an umber of `cargo-deny` failures,
stemming from an MPL copyleft license and an uncoordinated/differing
dependency between the `uefi-raw` and `uefi` packages.

This adds the licenses and the projects to the temporary `skip` list
for banning to prevent failures; but this will be fixed in the future.

Additionally, the various packages added recently have been given
licenses -- since this is necessary for `cargo-deny` to pass.