🩹 Remove `--workspace` from `cargo llvm-cov`
The `--workspace` flag when used with `cargo llvm-cov` builds
an entire workspace, even regardless of the default entries in the
workspace. This is resulting in the `bootloader` binary being built when
it should only ever be built for `*-uefi` targets.

The meaning of this flag was misunderstood when it was previously added
in the CI step, and so this has been removed to prevent further issues
and fix the broken CI.