✨ Add automatic release artifact creation
It can be useful for debugging to produce artifacts for every job in the
CD pipeline. Although this isn't a true "Deployment", it's producing an
artifact that can be used for deployment -- which can be helpful for
overall testing and ensuring consistency across builds.

This adds in a very basic "release" producedure that calls into a
`make build` command, which releases the `image` directory that is
built. This is not, yet, a real image (e.g. `*.iso`), but will be in the
future.

Tracked-by: #8