🔧 Generate CodeQL diagnostics from clippy
Github offers a `rust-clippy` workflow that is integrated with CodeQL
for code-scanning. Although this project currently has a `clippy`
invocation for static analysis, it does not release this as CodeQL --
which is actually a useful thing to be doing since it can provide
better feedback in the GUI for static analysis failures.

This adds in the artifact deployment for CodeQL code scanning.